requirements.txt
selenium==3.141.0
coloredlogs==15.0.1
line-bot-sdk==2.3.0
python-dotenv==0.21.0
